+++ nilearn/maskers/surface_maps_masker.py
@@ -13,11 +13,15 @@ from nilearn._utils import fill_doc, logger
 from nilearn._utils.cache_mixin import cache
 from nilearn._utils.class_inspect import get_params
 from nilearn.maskers._utils import (
-    check_surface_data_ndims,
     concat_extract_surface_data_parts,
 )
 from nilearn.maskers.base_masker import _BaseSurfaceMasker
-from nilearn.surface import SurfaceImage, check_same_n_vertices, concat_imgs
+from nilearn.surface.surface import (
+    SurfaceImage,
+    check_same_n_vertices,
+    check_surface_data_ndims,
+    concat_imgs,
+)
 
 
 @fill_doc
